+++ src/fides/api/service/connectors/consent_email_connector.py
@@ -103,6 +103,7 @@ class GenericConsentEmailConnector(BaseEmailConnector):
                                 preference=UserConsentPreference.opt_in,
                                 privacy_notice_history=PrivacyNoticeHistorySchema(
                                     name="Targeted Advertising",
+                                    notice_key="targeted_advertising",
                                     regions=["us_ca"],
                                     id="test_1",
                                     privacy_notice_id="12345",
@@ -117,6 +118,7 @@ class GenericConsentEmailConnector(BaseEmailConnector):
                                 preference=UserConsentPreference.opt_out,
                                 privacy_notice_history=PrivacyNoticeHistorySchema(
                                     name="Analytics",
+                                    notice_key="analytics",
                                     regions=["us_ca"],
                                     id="test_2",
                                     privacy_notice_id="67890",