+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The types of changes are:
 - Restrict strack-trace logging when not in Dev mode [#3081](https://github.com/ethyca/fides/pull/3081)
 - Refactor CSS variables for `fides-js` to match brandable color palette [#3321](https://github.com/ethyca/fides/pull/3321)
 - Moved all of the dirs from `fides.api.ops` into `fides.api` [#3318](https://github.com/ethyca/fides/pull/3318)
+- Add required notice key to privacy notices [#3337](https://github.com/ethyca/fides/pull/3337)
 
 ### Added